clean up before PR
