+++ garden_ai/gardens.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import logging
 from typing import TYPE_CHECKING, TypeVar
-from uuid import UUID
 from tabulate import tabulate
 
 from garden_ai.modal.functions import ModalFunction
@@ -257,61 +256,3 @@ class Garden:
         ]
 
         return cls(metadata, entrypoints, modal_functions, modal_classes)
-
-
-class AlphaFoldGarden(Garden):
-    """A Garden that uses a specific endpoint and function."""
-
-    def __init__(self, client, doi: str, endpoint_id: str, function_ids: list[str]):
-        self.client = client
-
-        # function_ids = ['bfc5557c-a1ff-4503-b276-8a8f6a4a4df1', '320aa08a-114e-4181-a393-98e3fcd02b0d']
-        function_names = ["predict", "check_prediction_status", "retrieve_results"]
-
-        stub_entrypoints = []
-        for function_id, function_name in zip(function_ids, function_names):
-            # Create the entrypoint first so we can use its DOI
-            entrypoint_metadata = RegisteredEntrypointMetadata(
-                doi=f"{doi}/{function_name}",
-                title="AlphaFold2 Prediction",
-                description="Predict protein structure using AlphaFold2",
-                short_name=function_name,
-                func_uuid=function_id,
-                container_uuid=UUID("00000000-0000-0000-0000-000000000000"),
-                base_image_uri="N/A",
-                full_image_uri="N/A",
-                notebook_url="https://thegardens.ai/",
-                function_text="",
-                entrypoint_ids=[],
-                doi_is_draft=False,
-            )
-            stub_entrypoints.append(Entrypoint(entrypoint_metadata))
-
-        metadata = GardenMetadata(
-            doi=doi,
-            title="AlphaFold2",
-            description="AlphaFold2 protein structure prediction",
-            entrypoint_ids=[ep.metadata.doi for ep in stub_entrypoints],
-            entrypoint_aliases={},
-            doi_is_draft=False,
-        )
-
-        super().__init__(metadata=metadata, entrypoints=stub_entrypoints)
-        self.endpoint_id = endpoint_id
-
-    # We'll call this 'submit' instead
-    def submit(self, fasta_string: str):
-        """Main prediction method that invokes the custom endpoint."""
-        print(f"Submitting HPC job to predict structure of {fasta_string}")
-        resp = self.entrypoints[0](fasta_string, endpoint=self.endpoint_id)
-        if type(resp) is dict:
-            return resp
-        raise ValueError(resp)
-
-    def check_prediction_status(self, task_id: str):
-        """Check the status of a submitted task."""
-        return self.entrypoints[1](task_id, endpoint=self.endpoint_id)
-
-    def retrieve_results(self, pdb_id: str):
-        """Retrieve the results of a submitted task."""
-        return self.entrypoints[2](pdb_id, endpoint=self.endpoint_id)
